Removed scroll button from mobile dimensions
by adding a custom hook that detects when window goes to mobile dimensions. The logic got removed from navbar to the custom hook
